+++ core/oslc4j-core/src/main/java/org/eclipse/lyo/oslc4j/core/model/ResourceShapeFactory.java
@@ -302,7 +302,7 @@ public class ResourceShapeFactory {
 
 		final OslcMaxSize maxSizeAnnotation = InheritedMethodAnnotationHelper.getAnnotation(method, OslcMaxSize.class);
 		if (maxSizeAnnotation != null) {
-			property.setMaxSize(maxSizeAnnotation.value());
+			property.setMaxSize(BigInteger.valueOf(maxSizeAnnotation.value()));
 		}
 
 		final OslcValueShape valueShapeAnnotation = InheritedMethodAnnotationHelper.getAnnotation(method, OslcValueShape.class);
